requirements.txt
flask
python-telegram-bot==20.6
transformers==4.36.2
torch==2.0.1
sentencepiece